Refactor repo, separate codes of direct-PC controlling and PC-PLC controlling
