+++ rga_etl/pc/scripts/rga_analog_scan.py
@@ -2,15 +2,15 @@ import os
 import datetime as dt
 from dotenv import load_dotenv
 from sqlalchemy.exc import IntegrityError
-from rga_etl.utils import init_session, init_instrument
-from rga_etl.rga import (
+from rga_etl.databases.utils import init_session, init_instrument
+from rga_etl.pc.rga import (
     init_rga,
     set_rga_analog_scan_parameters,
     set_rga_parameters_to_execution,
     rga_turn_off_filament,
 )
-from rga_etl.mysql import Execution, AnalogScan, AnalogScanPoint
-from rga_etl.fake import fake_analog_scan
+from rga_etl.databases.mysql import Execution, AnalogScan, AnalogScanPoint
+from rga_etl.pc.fake import fake_analog_scan
 
 
 def analog_scan(session):
